Struct forest_message::unsigned_message::UnsignedMessage

pub struct UnsignedMessage {
    pub version: i64,
    pub from: Address,
    pub to: Address,
    pub sequence: u64,
    pub value: TokenAmount,
    pub method_num: MethodNum,
    pub params: Serialized,
    pub gas_limit: i64,
    pub gas_fee_cap: TokenAmount,
    pub gas_premium: TokenAmount,
}

Default Unsigned VM message type which includes all data needed for a state transition

Usage:

use forest_message::{UnsignedMessage, Message};
use vm::{TokenAmount, Serialized, MethodNum};
use address::Address;

// Use the builder pattern to generate a message
let message = UnsignedMessage::builder()
    .to(Address::new_id(0))
    .from(Address::new_id(1))
    .sequence(0) // optional
    .value(TokenAmount::from(0u8)) // optional
    .method_num(MethodNum::default()) // optional
    .params(Serialized::default()) // optional
    .gas_limit(0) // optional
    .version(0) // optional
    .build()
    .unwrap();

let mut message_builder = UnsignedMessage::builder();
message_builder.sequence(1);
message_builder.from(Address::new_id(0));
message_builder.to(Address::new_id(1));
let msg = message_builder.build().unwrap();
assert_eq!(msg.sequence(), 1);

Fields
version: i64from: Addressto: Addresssequence: u64value: TokenAmountmethod_num: MethodNumparams: Serializedgas_limit: i64gas_fee_cap: TokenAmountgas_premium: TokenAmount
Implementations

impl UnsignedMessage

pub fn builder() -> MessageBuilder

pub fn to_signing_bytes(&self) -> Vec<u8>

Helper function to convert the message into signing bytes. This function returns the message Cid bytes.

pub fn valid_for_block_inclusion(
&self,
min_gas: i64,
version: NetworkVersion
) -> Result<(), String>

Semantic validation and validates the message has enough gas.